# docs/swagger.yaml
definitions:
  api_response.AuthTokens:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
      token_type:
        description: always "Bearer"
        type: string
    type: object
  api_response.CreateTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
  api_response.DeleteTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  api_response.DeleteUserResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  api_response.GetTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
  api_response.GetTodosResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/types.Pagination'
      status:
        type: integer
      success:
        type: boolean
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
    type: object
  api_response.GetUserResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
    type: object
  api_response.GetUsersResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/types.Pagination'
      status:
        type: integer
      success:
        type: boolean
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  api_response.LoginResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      tokens:
        $ref: '#/definitions/api_response.AuthTokens'
      user:
        $ref: '#/definitions/models.User'
    type: object
  api_response.LogoutResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  api_response.RefreshResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      tokens:
        $ref: '#/definitions/api_response.AuthTokens'
    type: object
  api_response.RegisterResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      tokens:
        $ref: '#/definitions/api_response.AuthTokens'
      user:
        $ref: '#/definitions/models.User'
    type: object
  api_response.UpdateTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.Todo:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_completed:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateTodoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      is_completed:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API documentation
  title: Example API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.LoginResponse'
      summary: Login
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.LogoutResponse'
      summary: Logout (revoke refresh token)
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.RefreshResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_response.RegisterResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/todos:
    get:
      description: Retrieve todos with optional pagination (page, size) and filter
        by is_completed
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: size
        type: integer
      - description: Filter by completion status
        in: query
        name: is_completed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTodosResponse'
      summary: Get todos with pagination and filters
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.CreateTodoResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
      summary: Create a new todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      description: Delete a specific todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.DeleteTodoResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Delete todo by ID
      tags:
      - todos
    get:
      description: Fetch a specific todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTodoResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Get todo by ID
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update a todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateTodoResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Update an existing todo
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update a todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateTodoResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      summary: Update an existing todo
      tags:
      - todos
  /api/v1/users:
    get:
      description: Retrieve users with optional pagination (page, size)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetUsersResponse'
      security:
      - BearerAuth: []
      summary: Get users with pagination and filters
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Delete a specific user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.DeleteUserResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - users
    get:
      description: Fetch a specific user by its ID
      parameters:
      - description: user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetUserResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"